Revert "move statusCodeAndMessage_withLatchedCallExecutor out of AbstractInteropTest"
This reverts commit d2c7049095271a63365f48549d0be80a0d7eeae3.
