+++ interop-testing/src/test/java/io/grpc/testing/integration/AutoWindowSizingOnTest.java
@@ -27,7 +27,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
-public class AutoWindowSizingOnTest extends MoreInteropTests {
+public class AutoWindowSizingOnTest extends AbstractInteropTest {
 
   @BeforeClass
   public static void turnOnAutoWindow() {
