+++ interop-testing/src/test/java/io/grpc/testing/integration/Http2OkHttpTest.java
@@ -52,7 +52,7 @@ import org.junit.runners.JUnit4;
  * Integration tests for GRPC over Http2 using the OkHttp framework.
  */
 @RunWith(JUnit4.class)
-public class Http2OkHttpTest extends MoreInteropTests {
+public class Http2OkHttpTest extends AbstractInteropTest {
 
   private static final String BAD_HOSTNAME = "I.am.a.bad.hostname";
